Jan 2022 Updates

+++ lib/opt/trash.sh
@@ -14,30 +14,87 @@ else
     mkdir -p "$OMS_CACHE/trash"
 fi
 
+# Trash-Auto-Deleting Service
+
+if [ -n "$trashAutoDeleteService" ] && [ "$trashAutoDeleteService" = "Enable" ]
+then
+    find "$OMS_CACHE/trash" -maxdepth 1 -printf "%f\n" | while IFS= read -r i
+    do
+        fullfilename="${i//"_%^%_"/"/"}"
+        filename="${fullfilename%%"_%backup%_"*}"
+        timename="${fullfilename##*"_%backup%_"}"
+        if [ "$timename" != "$filename" ]
+        then
+            editdate="${timename:0:4}/${timename:4:2}/${timename:6:2} ${timename:9:2}:${timename:11:2}:${timename:13:2}.${timename:16:8}"
+            secdate="$(date -d "$editdate" "+%s")"
+            time1=$((($(date +%s)-secdate)/86400))
+            if [ "$time1" -gt "$trashAutoDeleteConfigDate" ]
+            then
+                _warn "The file $filename last over $trashAutoDeleteConfigDate days, deleting..."
+                rm -rf "$OMS_CACHE/trash/$i"
+            fi
+        fi
+    done
+fi
+
+helptrash(){
+    cat <<EOF
+              trash help:
+trash:
+  trash FILE       :  move a file/folder to the bin
+lstrash:
+  lstrash          :  get a list of the bin
+  lstrash KEY      :  search KEY
+  lstrash PATH     :  search file on PATH
+rmtrash:
+  rmtrash FILE     :  remove FILE
+  rmtrash KEY      :  search KEY to remove
+  rmtrash -a       :  remove all files
+retrash:
+  retrash FILE     :  restore FILE
+  retrash KEY      :  search KEY to restore
+EOF
+}
+
 trash(){
     if [ -z "$1" ]
     then
         _error 'Missing parameters' 'OhMySh::TrashManager' '10'
     else
-        RP="$(realpath "$1")"
-        NP="${RP//"/"/"_%^%_"}"
-        _info "Deleting '$RP'"
-        # if [ -f "$OMS_CACHE/trash/$NP" ] || [ -d "$OMS_CACHE/trash/$NP" ]
-        # then
-            # mv "$OMS_CACHE/trash/$NP" "$OMS_CACHE/trash/${NP}_%backup%_$(date +%Y%m%d_%H%M%S_%8N)"
-        # fi    
-        # mv "$RP" "$OMS_CACHE/trash/$NP"
-        mv "$RP" "$OMS_CACHE/trash/${NP}_%backup%_$(date +%Y%m%d_%H%M%S_%8N)"
+        if [ -f "$1" ] || [ -d "$1" ]
+        then
+            RP="$(realpath "$1")"
+            NP="${RP//"/"/"_%^%_"}"
+            NP="${NP//" "/"_%SPACE%_"}"
+            _info "Deleting '$RP'"
+            # echo $NP
+            # if [ -f "$OMS_CACHE/trash/$NP" ] || [ -d "$OMS_CACHE/trash/$NP" ]
+            # then
+                # mv "$OMS_CACHE/trash/$NP" "$OMS_CACHE/trash/${NP}_%backup%_$(date +%Y%m%d_%H%M%S_%8N)"
+            # fi
+            # mv "$RP" "$OMS_CACHE/trash/$NP"
+            mv "$RP" "$OMS_CACHE/trash/${NP}_%backup%_$(date +%Y%m%d_%H%M%S_%8N)"
+        else
+            _error "No such file or directory." "TrashManager"
+        fi
     fi
 }
 
 lstrash(){
     # ls "$OMS_CACHE/trash"
+    if [ -n "$1" ]
+    then
+        search_1="$(realpath "$1")"
+        search="${search_1//"/"/"_%^%_"}"
+    else
+        search=""
+    fi
     printf " %-22s %s\n" "Date deleted" "File name"
-    find "$OMS_CACHE/trash" -maxdepth 1 -printf "%f\n" | while IFS= read -r i
+    find "$OMS_CACHE/trash" -maxdepth 1 -printf "%f\n" | grep "$search" | while IFS= read -r i
     do
         fullfilename="${i//"_%^%_"/"/"}"
         # echo $fullfilename
+        fullfilename="${fullfilename//"_%SPACE%_"/" "}"
         filename="${fullfilename%%"_%backup%_"*}"
         timename="${fullfilename##*"_%backup%_"}"
         # echo "$timename"
@@ -65,14 +122,14 @@ lstrash(){
 }
 
 rmtrash(){
-    if [ -z "$1" ]
-    then
-        _error 'Missing parameters' 'OhMySh::TrashManager' '10'
-    else
+#     if [ -z "$1" ]
+#     then
+#         _error 'Missing parameters' 'OhMySh::TrashManager' '10'
+#     else
         if [ "$1" = "-a" ]
         then
             _warn "You are going to delete all the caching trashes. Are you sure? [y/n]" "TrashManager"
-            read -s -n1 _cfm
+            read -r -s -n1 _cfm
             if [ "$_cfm" = "y" ] || [ "$_cfm" = "Y" ]
             then
                 _warn "You chose yes." "TrashManager"
@@ -81,7 +138,7 @@ rmtrash(){
                 blue "You chose no. Nothing will be changed."
             fi
         else
-            RP="$(realpath "$1")"
+            RP="$1"
             NP="${RP//"/"/"_%^%_"}"
             _warn "Deleting $RP..."
 #             echo $NP
@@ -106,40 +163,45 @@ rmtrash(){
                 for ((i=0;i<_count;i++))
                 do
                     fullfilename="${_rmlist[i]//"_%^%_"/"/"}"
+                    fullfilename="${fullfilename//"_%SPACE%_"/" "}"
                     filename="${fullfilename%%"_%backup%_"*}"
                     timename="${fullfilename##*"_%backup%_"}"
-                    editdate="${timename:0:4}/${timename:4:2}/${timename:6:2} ${timename:9:2}:${timename:11:2}:${timename:13:2}.${timename:16:8}"
-                    outdate="$(date -d "$editdate" "+%F %H:%M:%S")"
-                    if [ -d "$OMS_CACHE/trash/${_rmlist[i]}" ]
+                    if [ "$timename" != "$filename" ]
                     then
-                        printf " %-4s %-22s \033[34m%s\033[0m\n" "$i" "$outdate" "$filename"
-                    elif [ -f "$OMS_CACHE/trash/${_rmlist[i]}" ]
-                    then
-                        printf " %-4s %-22s %s\n" "$i" "$outdate" "$filename"
+                        editdate="${timename:0:4}/${timename:4:2}/${timename:6:2} ${timename:9:2}:${timename:11:2}:${timename:13:2}.${timename:16:8}"
+                        outdate="$(date -d "$editdate" "+%F %H:%M:%S")"
+                        if [ -d "$OMS_CACHE/trash/${_rmlist[i]}" ]
+                        then
+                            printf " %-4s %-22s \033[34m%s\033[0m\n" "$i" "$outdate" "$filename"
+                        elif [ -f "$OMS_CACHE/trash/${_rmlist[i]}" ]
+                        then
+                            printf " %-4s %-22s %s\n" "$i" "$outdate" "$filename"
+                        fi
                     fi
                 done
                 _info "Which one do you want to delete? [0-$((_count-1))]"
-                read _delnum
+                read -r _delnum
                 if [ -z "$_delnum" ] || [ "$_delnum" -lt "0" ] || [ "$_delnum" -ge "$_count" ]
                 then
                     _error "Invalid input."
                     return
                 fi
                 fullfilename="${_rmlist[_delnum]//"_%^%_"/"/"}"
+                fullfilename="${fullfilename//"_%SPACE%_"/" "}"
                 filename="${fullfilename%%"_%backup%_"*}"
                 _warn "Deleting ${filename}..."
-                rm -rf "${_rmlist[_delnum]}"
+                rm -rf "$OMS_CACHE/trash/${_rmlist[_delnum]}"
             fi
         fi
-    fi
+#     fi
 }
 
 retrash(){
-    if [ -z "$1" ]
-    then
-        _error 'Missing parameters' 'OhMySh::TrashManager' '10'
-    else
-        RP="$(realpath "$1")"
+#     if [ -z "$1" ]
+#     then
+#         _error 'Missing parameters' 'OhMySh::TrashManager' '10'
+#     else
+        RP="$1"
         NP="${RP//"/"/"_%^%_"}"
         _warn "Restoring $RP..."
 #         _rmlist=()
@@ -161,29 +223,34 @@ retrash(){
             for ((i=0;i<_count;i++))
             do
                 fullfilename="${_rmlist[i]//"_%^%_"/"/"}"
+                fullfilename="${fullfilename//"_%SPACE%_"/" "}"
                 filename="${fullfilename%%"_%backup%_"*}"
                 timename="${fullfilename##*"_%backup%_"}"
-                editdate="${timename:0:4}/${timename:4:2}/${timename:6:2} ${timename:9:2}:${timename:11:2}:${timename:13:2}.${timename:16:8}"
-                outdate="$(date -d "$editdate" "+%F %H:%M:%S")"
-                if [ -d "$OMS_CACHE/trash/${_rmlist[i]}" ]
+                if [ "$timename" != "$filename" ]
                 then
-                    printf " %-4s %-22s \033[34m%s\033[0m\n" "$i" "$outdate" "$filename"
-                elif [ -f "$OMS_CACHE/trash/${_rmlist[i]}" ]
-                then
-                    printf " %-4s %-22s %s\n" "$i" "$outdate" "$filename"
+                    editdate="${timename:0:4}/${timename:4:2}/${timename:6:2} ${timename:9:2}:${timename:11:2}:${timename:13:2}.${timename:16:8}"
+                    outdate="$(date -d "$editdate" "+%F %H:%M:%S")"
+                    if [ -d "$OMS_CACHE/trash/${_rmlist[i]}" ]
+                    then
+                        printf " %-4s %-22s \033[34m%s\033[0m\n" "$i" "$outdate" "$filename"
+                    elif [ -f "$OMS_CACHE/trash/${_rmlist[i]}" ]
+                    then
+                        printf " %-4s %-22s %s\n" "$i" "$outdate" "$filename"
+                    fi
                 fi
             done
             _info "Which one do you want to delete? [0-$((_count-1))]"
-            read _delnum
+            read -r _delnum
             if [ -z "$_delnum" ] || [ "$_delnum" -lt "0" ] || [ "$_delnum" -ge "$_count" ]
             then
                 _error "Invalid input."
                 return
             fi
             fullfilename="${_rmlist[_delnum]//"_%^%_"/"/"}"
+            fullfilename="${fullfilename//"_%SPACE%_"/" "}"
             filename="${fullfilename%%"_%backup%_"*}"
             _warn "Restoring ${filename}..."
-            mv -r "${_rmlist[_delnum]}" "$filename"
+            mv -f "$OMS_CACHE/trash/${_rmlist[_delnum]}" "$filename"
         fi
-    fi
+#     fi
 }
